+++ toolkit/sharedlib/src/main/java/com/arcgismaps/toolkit/ui/expandablecard/ExpandableCard.kt
@@ -50,7 +50,6 @@ import com.arcgismaps.toolkit.ui.expandablecard.theme.ExpandableCardDefaults
 import com.arcgismaps.toolkit.ui.expandablecard.theme.ExpandableCardShapes
 import com.arcgismaps.toolkit.ui.expandablecard.theme.ExpandableCardTheme
 import com.arcgismaps.toolkit.ui.expandablecard.theme.ExpandableCardTypography
-import com.arcgismaps.toolkit.ui.modifier.applyIf
 
 /**
  * Composable Card that has the ability to expand and collapse its [content].
@@ -124,12 +123,8 @@ private fun ExpandableHeader(
     if (title.isEmpty() && description == null && !expandable) return
     Row(
         Modifier
+            .then(if(expandable) Modifier.clickable { onClick() } else Modifier)
             .fillMaxWidth()
-            .applyIf(expandable) {
-                clickable {
-                    onClick()
-                }
-            }
             .background(colors.headerBackgroundColor),
         verticalAlignment = Alignment.CenterVertically
     ) {
@@ -166,7 +161,7 @@ private fun ExpandableHeader(
 
 @Preview
 @Composable
-internal fun ExpandableHeaderPreview() {
+private fun ExpandableHeaderPreview() {
     ExpandableHeader(
         title = "The Title",
         colors = ExpandableCardDefaults.colorScheme(),
